% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iopt-package.R
\docType{package}
\name{ipopt-package}
\alias{ipopt}
\alias{ipopt-package}
\title{Ipopt package library from COIN OR tools which can solve nonlinear optimization problems}
\description{
Provides a lightweight R interface to the Ipopt nonlinear optimization solver through Ipopt's C interface. The package is intended for r-universe builds that provide Ipopt and its open linear solver dependencies, such as MUMPS, through system libraries discoverable by pkg-config.
}
\author{
\strong{Maintainer}: Balasubramanian Narasimhan \email{naras@stanford.edu}

Other contributors:
\itemize{
  \item Andreas Waechter (Main author of Ipopt; COIN-OR project manager) [contributor, copyright holder]
  \item Carl Laird (Main author of Ipopt) [contributor, copyright holder]
  \item Stefan Vigerske (COIN-OR project manager for Ipopt) [contributor]
  \item COIN-OR Foundation (Ipopt project and copyright holder) [copyright holder]
}

}
\keyword{internal}
