% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipopt.R
\name{ipopt_version}
\alias{ipopt_version}
\title{Ipopt Library Version}
\usage{
ipopt_version()
}
\value{
An integer vector with elements \code{major}, \code{minor}, and \code{release}.
}
\description{
Ipopt Library Version
}
